@@ -3,7 +3,7 @@ package lucas.prices.reposotory;
 
 import lombok.Data;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 @Data
